Move `rustc_ast::AutoDiffAttrs` to `rustc_hir::RustcAutodiff`
