+++ compiler/rustc_ast/src/expand/autodiff_attrs.rs
@@ -1,12 +1,13 @@
 //! This crate handles the user facing autodiff macro. For each `#[autodiff(...)]` attribute,
-//! we create an [`AutoDiffItem`] which contains the source and target function names. The source
+//! we create an `RustcAutodiff` which contains the source and target function names. The source
 //! is the function to which the autodiff attribute is applied, and the target is the function
 //! getting generated by us (with a name given by the user as the first autodiff arg).
 
 use std::fmt::{self, Display, Formatter};
 use std::str::FromStr;
 
-use crate::expand::typetree::TypeTree;
+use rustc_span::{Symbol, sym};
+
 use crate::expand::{Decodable, Encodable, HashStable_Generic};
 use crate::{Ty, TyKind};
 
@@ -29,6 +30,12 @@ pub enum DiffMode {
     Forward,
     /// The target function, to be created using reverse mode AD.
     Reverse,
+}
+
+impl DiffMode {
+    pub fn all_modes() -> &'static [Symbol] {
+        &[sym::Source, sym::Forward, sym::Reverse]
+    }
 }
 
 /// Dual and Duplicated (and their Only variants) are getting lowered to the same Enzyme Activity.
@@ -76,43 +83,20 @@ impl DiffActivity {
         use DiffActivity::*;
         matches!(self, |Dual| DualOnly | Dualv | DualvOnly | Const)
     }
-}
-/// We generate one of these structs for each `#[autodiff(...)]` attribute.
-#[derive(Clone, Eq, PartialEq, Encodable, Decodable, Debug, HashStable_Generic)]
-pub struct AutoDiffItem {
-    /// The name of the function getting differentiated
-    pub source: String,
-    /// The name of the function being generated
-    pub target: String,
-    pub attrs: AutoDiffAttrs,
-    pub inputs: Vec<TypeTree>,
-    pub output: TypeTree,
-}
 
-#[derive(Clone, Eq, PartialEq, Encodable, Decodable, Debug, HashStable_Generic)]
-pub struct AutoDiffAttrs {
-    /// Conceptually either forward or reverse mode AD, as described in various autodiff papers and
-    /// e.g. in the [JAX
-    /// Documentation](https://jax.readthedocs.io/en/latest/_tutorials/advanced-autodiff.html#how-it-s-made-two-foundational-autodiff-functions).
-    pub mode: DiffMode,
-    /// A user-provided, batching width. If not given, we will default to 1 (no batching).
-    /// Calling a differentiated, non-batched function through a loop 100 times is equivalent to:
-    /// - Calling the function 50 times with a batch size of 2
-    /// - Calling the function 25 times with a batch size of 4,
-    /// etc. A batched function takes more (or longer) arguments, and might be able to benefit from
-    /// cache locality, better re-usal of primal values, and other optimizations.
-    /// We will (before LLVM's vectorizer runs) just generate most LLVM-IR instructions `width`
-    /// times, so this massively increases code size. As such, values like 1024 are unlikely to
-    /// work. We should consider limiting this to u8 or u16, but will leave it at u32 for
-    /// experiments for now and focus on documenting the implications of a large width.
-    pub width: u32,
-    pub ret_activity: DiffActivity,
-    pub input_activity: Vec<DiffActivity>,
-}
-
-impl AutoDiffAttrs {
-    pub fn has_primal_ret(&self) -> bool {
-        matches!(self.ret_activity, DiffActivity::Active | DiffActivity::Dual)
+    pub fn all_activities() -> &'static [Symbol] {
+        &[
+            sym::None,
+            sym::Active,
+            sym::ActiveOnly,
+            sym::Const,
+            sym::Dual,
+            sym::Dualv,
+            sym::DualOnly,
+            sym::DualvOnly,
+            sym::Duplicated,
+            sym::DuplicatedOnly,
+        ]
     }
 }
 
@@ -239,61 +223,5 @@ impl FromStr for DiffActivity {
             "DuplicatedOnly" => Ok(DiffActivity::DuplicatedOnly),
             _ => Err(()),
         }
-    }
-}
-
-impl AutoDiffAttrs {
-    pub fn has_ret_activity(&self) -> bool {
-        self.ret_activity != DiffActivity::None
-    }
-    pub fn has_active_only_ret(&self) -> bool {
-        self.ret_activity == DiffActivity::ActiveOnly
-    }
-
-    pub const fn error() -> Self {
-        AutoDiffAttrs {
-            mode: DiffMode::Error,
-            width: 0,
-            ret_activity: DiffActivity::None,
-            input_activity: Vec::new(),
-        }
-    }
-    pub fn source() -> Self {
-        AutoDiffAttrs {
-            mode: DiffMode::Source,
-            width: 0,
-            ret_activity: DiffActivity::None,
-            input_activity: Vec::new(),
-        }
-    }
-
-    pub fn is_active(&self) -> bool {
-        self.mode != DiffMode::Error
-    }
-
-    pub fn is_source(&self) -> bool {
-        self.mode == DiffMode::Source
-    }
-    pub fn apply_autodiff(&self) -> bool {
-        !matches!(self.mode, DiffMode::Error | DiffMode::Source)
-    }
-
-    pub fn into_item(
-        self,
-        source: String,
-        target: String,
-        inputs: Vec<TypeTree>,
-        output: TypeTree,
-    ) -> AutoDiffItem {
-        AutoDiffItem { source, target, inputs, output, attrs: self }
-    }
-}
-
-impl fmt::Display for AutoDiffItem {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "Differentiating {} -> {}", self.source, self.target)?;
-        write!(f, " with attributes: {:?}", self.attrs)?;
-        write!(f, " with inputs: {:?}", self.inputs)?;
-        write!(f, " with output: {:?}", self.output)
     }
 }
