+++ compiler/rustc_builtin_macros/src/autodiff.rs
@@ -8,8 +8,7 @@ mod llvm_enzyme {
     use std::string::String;
 
     use rustc_ast::expand::autodiff_attrs::{
-        AutoDiffAttrs, DiffActivity, DiffMode, valid_input_activity, valid_ret_activity,
-        valid_ty_for_activity,
+        DiffActivity, DiffMode, valid_input_activity, valid_ret_activity, valid_ty_for_activity,
     };
     use rustc_ast::token::{Lit, LitKind, Token, TokenKind};
     use rustc_ast::tokenstream::*;
@@ -20,6 +19,7 @@ mod llvm_enzyme {
         MetaItemInner, MgcaDisambiguation, PatKind, Path, PathSegment, TyKind, Visibility,
     };
     use rustc_expand::base::{Annotatable, ExtCtxt};
+    use rustc_hir::attrs::RustcAutodiff;
     use rustc_span::{Ident, Span, Symbol, sym};
     use thin_vec::{ThinVec, thin_vec};
     use tracing::{debug, trace};
@@ -87,7 +87,7 @@ mod llvm_enzyme {
         meta_item: &ThinVec<MetaItemInner>,
         has_ret: bool,
         mode: DiffMode,
-    ) -> AutoDiffAttrs {
+    ) -> RustcAutodiff {
         let dcx = ecx.sess.dcx();
 
         // Now we check, whether the user wants autodiff in batch/vector mode, or scalar mode.
@@ -105,7 +105,7 @@ mod llvm_enzyme {
                         span: meta_item[1].span(),
                         width: x,
                     });
-                    return AutoDiffAttrs::error();
+                    return RustcAutodiff::error();
                 }
             }
         } else {
@@ -129,7 +129,7 @@ mod llvm_enzyme {
             };
         }
         if errors {
-            return AutoDiffAttrs::error();
+            return RustcAutodiff::error();
         }
 
         // If a return type exist, we need to split the last activity,
@@ -145,11 +145,11 @@ mod llvm_enzyme {
             (&DiffActivity::None, activities.as_slice())
         };
 
-        AutoDiffAttrs {
+        RustcAutodiff {
             mode,
             width,
             ret_activity: *ret_activity,
-            input_activity: input_activity.to_vec(),
+            input_activity: input_activity.iter().cloned().collect(),
         }
     }
 
@@ -309,7 +309,7 @@ mod llvm_enzyme {
         ts.pop();
         let ts: TokenStream = TokenStream::from_iter(ts);
 
-        let x: AutoDiffAttrs = from_ast(ecx, &meta_item_vec, has_ret, mode);
+        let x: RustcAutodiff = from_ast(ecx, &meta_item_vec, has_ret, mode);
         if !x.is_active() {
             // We encountered an error, so we return the original item.
             // This allows us to potentially parse other attributes.
@@ -603,7 +603,7 @@ mod llvm_enzyme {
     fn gen_enzyme_decl(
         ecx: &ExtCtxt<'_>,
         sig: &ast::FnSig,
-        x: &AutoDiffAttrs,
+        x: &RustcAutodiff,
         span: Span,
     ) -> ast::FnSig {
         let dcx = ecx.sess.dcx();
